try4

+++ scene/gaussian_model.py
@@ -499,7 +499,7 @@ class GaussianModel:
         # max variance = max eigenvalue of covariance matrix
         covariances = self.get_actual_covariances()
         eigvals = torch.linalg.eigvals(covariances)
-        v = torch.max(eigvals)
+        v = torch.max(eigvals, dtype = torch.float)
         return v / (1 + torch.exp(- self.growth_length_s[0]))
     
     def get_actual_covariances (self, scaling_modifier = 1):
